# `Keccak-p[1600, 24]` Permutation using AVX2 Intrinsics