--- a/Makefile
+++ b/Makefile
@@ -24,7 +24,7 @@ clean:
 	find . -name '*.out' -o -name '*.o' -o -name '*.so' -o -name '*.gch' | xargs rm -rf
 
 format:
-	find . -name '*.cpp' -o -name '*.hpp' | xargs clang-format -i --style="{BasedOnStyle: Mozilla, ColumnLimit: 100}" && python3 -m black wrapper/python/*.py
+	find . -name '*.cpp' -o -name '*.hpp' | xargs clang-format -i --style="{BasedOnStyle: Mozilla, ColumnLimit: 110}" && python3 -m black wrapper/python/*.py
 
 bench/a.out: bench/main.cpp include/*.hpp
 	# make sure you've google-benchmark globally installed;