Added optional compilation using AVX2 intrinsics on platforms supporting AVX2. 

> **Note**
Define `USE_AVX2` flag at compile-time.

## On Intel(R) Core(TM) i5-8279U CPU @ 2.40GHz

When issued `make benchmark`

### Without AVX2

```bash
2022-12-05T10:06:19+04:00
Running ./bench/a.out
Run on (8 X 2400 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB
  L1 Instruction 32 KiB
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB
Load Average: 1.67, 1.80, 2.01
------------------------------------------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations average_cpu_cycles bytes_per_second
------------------------------------------------------------------------------------------------------------
bench_sha3::keccakf1600              354 ns          354 ns      2000177                831       538.831M/s
```

### With AVX2

```bash
2022-12-05T10:07:16+04:00
Running ./bench/a.out
Run on (8 X 2400 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB
  L1 Instruction 32 KiB
  L2 Unified 256 KiB (x4)
  L3 Unified 6144 KiB
Load Average: 2.03, 1.88, 2.02
------------------------------------------------------------------------------------------------------------
Benchmark                              Time             CPU   Iterations average_cpu_cycles bytes_per_second
------------------------------------------------------------------------------------------------------------
bench_sha3::keccakf1600              387 ns          386 ns      1837714                911       493.699M/s
```

Looks like AVX2 is not helping much here !