@itzmeanjan 2022-12-05
At commit https://github.com/itzmeanjan/sha3/pull/16/commits/4a6149760218e4900eb279335fed7a7d5cc65abe, on AWS Graviton2 ( aarch64 architecture ), when profiled with `perf`

<img width="830" alt="image" src="https://user-images.githubusercontent.com/45074836/205703551-b1142dd7-c029-4b70-940c-b8956c242365.png">


At commit https://github.com/itzmeanjan/sha3/pull/16/commits/68bea202191783b8797301b17c6fc88704186cc6, on AWS Graviton2 ( aarch64 architecture ), when profiled with `perf`

<img width="830" alt="image" src="https://user-images.githubusercontent.com/45074836/205703684-f70a1a58-4d8a-4a4d-ad8c-8eefadaba6bd.png">

> **Note**
See how reducing branch misprediction rate by unrolling the loops and using precomputed permutation table, can boost performance of keccak-p[1600, 24] permutation on ARM-A profile CPU.