@itzmeanjan 2022-12-05
At commit https://github.com/itzmeanjan/sha3/pull/16/commits/4a6149760218e4900eb279335fed7a7d5cc65abe, on AWS Graviton3 ( aarch64 architecture, ARM Neoverse core ), when profiled with `perf` [ **BEFORE** ]

<img width="830" alt="image" src="https://user-images.githubusercontent.com/45074836/205704261-d2a51605-e49b-426d-85ab-71f71001ddc0.png">


At commit https://github.com/itzmeanjan/sha3/pull/16/commits/68bea202191783b8797301b17c6fc88704186cc6, on AWS Graviton3 ( aarch64 architecture, ARM Neoverse core ), when profiled with `perf` [ **AFTER** ]

<img width="830" alt="image" src="https://user-images.githubusercontent.com/45074836/205704291-6f687ab6-b384-4b90-88d7-bf23abf6520e.png">

> **Note**
How branch misprediction rate drops from ~0.5256% to ~0.04% and brings huge performance improvement for Keccak-p[1600, 24] permutation.